add example object detection with ui
