+++ examples-basics/object-detection/04-object-detection-with-ui/assets/docs_assets/thumbnail.png
