+++ examples-basics/object-detection/04-object-detection-with-ui/assets/fonts/Open Sans/OFL.txt
