+++ examples-basics/object-detection/04-object-detection-with-ui/assets/img/favicon.png
